refactor(answerView): modularize heuristic answer component
